Completed Toolbar breakup into seperate components.
