+++ src/__testData__/fullRenderedComponent.js
@@ -1,7 +1,7 @@
 import Browse from "../SharedComponents/Browse/Browse";
 import MyWorks from "../Author/MyWorks/MyWorks";
 import Dashboard from "../SharedComponents/Dashboard/Dashboard";
-import Footer from "../SharedComponents/Footer";
+import Footer from "../SharedComponents/Footer/Footer";
 import Header from "../SharedComponents/Header";
 import Profile from "../Author/Profile/Profile.jsx";
 import UploadModal from "../Author/MyWorks/UploadModal.jsx";
